requirements.txt
google-generativeai
pymupdf
pytesseract
pillow
pdf2image
langgraph
crewai
fastapi
uvicorn
gradio
pydantic
python-multipart
requests